+++ src/mig/scheduler/context.go
@@ -55,7 +55,7 @@ import (
 // database and message brokers. It also contains some statistics.
 // Context is intended as a single structure that can be passed around easily.
 type Context struct {
-	OpID  uint64 // ID of the current operation, used for tracking
+	OpID  float64 // ID of the current operation, used for tracking
 	Agent struct {
 		// configuration
 		TimeOut, HeartbeatFreq, Whitelist string
